Bump Sphinx version, improve README.md instructions
Sphinx 3.5.3 has a bug that causes the addition of a `custom.css` file to break
the base `theme.css` rules.  Upgrading to Sphinx 3.5.4 fixes this issue.

Additionally, if the user has `sphinx-build` or `sphinx-autobuild` installed
elsewhere on their PYTHONPATH and forgets to activate the Python virtual
environment, strange things can happen.

To avoid this issue, I have changed `README.md` as follows:

- Prefix all virtual environment commands with `_venv/bin` to ensure the
  correct executables are run

- Drop the recommendation to use the `activate` script